@@ -16,18 +16,19 @@
     "yarn": ">=1.22.0"
   },
   "dependencies": {
-    "@ls1intum/apollon": "2.12.9",
+    "@ls1intum/apollon": "2.13.0",
     "bootstrap": "5.2.3",
     "moment": "2.29.4",
-    "react": "17.0.2",
+    "postcss-syntax": "0.36.2",
+    "react": "18.2.0",
     "react-bootstrap": "2.6.0",
     "react-bootstrap-icons": "^1.10.2",
-    "react-dom": "17.0.2",
-    "react-redux": "7.2.6",
-    "react-router-dom": "5.3.0",
+    "react-dom": "18.2.0",
+    "react-redux": "8.0.5",
+    "react-router-dom": "6.8.1",
     "redux": "4.2.0",
     "redux-observable": "2.0.0",
-    "rxjs": "7.3.0",
+    "rxjs": "7.8.0",
     "shared": "1.6.9",
     "styled-components": "5.3.6",
     "tslib": "2.4.1",
@@ -36,12 +37,12 @@
     "websocket": "1.0.34"
   },
   "devDependencies": {
+    "@stylelint/postcss-css-in-js": "0.38.0",
     "@svgr/webpack": "^6.5.1",
-    "@types/react": "17.0.37",
+    "@types/react": "18.0.27",
     "@types/react-bootstrap": "0.32.32",
-    "@types/react-dom": "17.0.11",
-    "@types/react-redux": "7.1.24",
-    "@types/react-router-dom": "5.3.3",
+    "@types/react-dom": "18.0.10",
+    "@types/react-redux": "7.1.25",
     "@types/styled-components": "5.1.26",
     "@types/uuid": "9.0.0",
     "circular-dependency-plugin": "5.2.2",
@@ -51,11 +52,12 @@
     "fork-ts-checker-notifier-webpack-plugin": "6.0.0",
     "fork-ts-checker-webpack-plugin": "7.2.13",
     "html-webpack-plugin": "5.5.0",
+    "postcss": "8.4.21",
     "prettier": "2.8.0",
-    "react-toastify": "^9.1.1",
+    "react-toastify": "9.1.1",
     "style-loader": "3.3.1",
-    "stylelint": "14.16.0",
-    "stylelint-config-recommended": "9.0.0",
+    "stylelint": "15.1.0",
+    "stylelint-config-recommended": "10.0.1",
     "stylelint-config-styled-components": "0.1.1",
     "stylelint-custom-processor-loader": "0.6.0",
     "stylelint-processor-styled-components": "1.10.0",
