@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { ApplicationBar } from './components/application-bar/application-bar';
 import { ApollonEditorWrapper } from './components/apollon-editor-component/apollon-editor-component';
 import { ApollonEditor, ApollonOptions } from '@ls1intum/apollon';
@@ -21,23 +21,9 @@ import { defaultEditorOptions } from './services/editor-options/editor-options-r
 import { EditorOptions } from './services/editor-options/editor-options-types';
 import { ErrorPanel } from './components/error-handling/error-panel';
 import { Diagram } from './services/diagram/diagram-types';
-import { BrowserRouter, RouteComponentProps, Route, Switch } from 'react-router-dom';
+import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import { ApplicationModal } from './components/modals/application-modal';
 import { ToastContainer } from 'react-toastify';
-
-type OwnProps = {};
-
-type Props = OwnProps & RouteComponentProps;
-
-type State = {
-  options?: ApollonOptions;
-  editor?: ApollonEditor;
-};
-
-const initialState: State = Object.freeze({
-  model: undefined as ApollonOptions | undefined,
-  editor: undefined as ApollonEditor | undefined,
-});
 
 const getInitialStore = (): ApplicationState => {
   const latestId: string | null = window.localStorage.getItem(localStorageLatest);
@@ -73,39 +59,37 @@ const getInitialStore = (): ApplicationState => {
 
 const initialStore = getInitialStore();
 
-export class Application extends React.Component<Props, State> {
-  state = initialState;
-  setEditor = (ref: ApollonEditor) => {
+export const Application = () => {
+  const [editor, setEditor] = useState<ApollonEditor>();
+  const handleSetEditor = (ref: ApollonEditor) => {
     if (ref) {
-      this.setState({ editor: ref });
+      setEditor(ref);
     }
   };
+  const isFirefox: boolean = /Firefox/i.test(navigator.userAgent);
+  const context: ApollonEditorContext | null = { editor, setEditor: handleSetEditor };
 
-  render() {
-    const isFirefox: boolean = /Firefox/i.test(navigator.userAgent);
-    const context: ApollonEditorContext | null = { editor: this.state.editor, setEditor: this.setEditor };
-    return (
-      <ApollonEditorProvider value={context}>
-        <ApplicationStore initialState={initialStore}>
-          <ApplicationBar />
-          <ApplicationModal />
-          {isFirefox && <FirefoxIncompatibilityHint />}
-          <ErrorPanel />
-          <ApollonEditorWrapper />
-        </ApplicationStore>
-        <ToastContainer />
-      </ApollonEditorProvider>
-    );
-  }
-}
+  return (
+    <ApollonEditorProvider value={context}>
+      <ApplicationStore initialState={initialStore}>
+        <ApplicationBar />
+        <ApplicationModal />
+        {isFirefox && <FirefoxIncompatibilityHint />}
+        <ErrorPanel />
+        <ApollonEditorWrapper />
+      </ApplicationStore>
+      <ToastContainer />
+    </ApollonEditorProvider>
+  );
+};
 
 export function RoutedApplication() {
   return (
     <BrowserRouter>
-      <Switch>
-        <Route path={'/:token'} render={(props) => <Application {...props} />} />
-        <Route path={'/'} render={(props) => <Application {...props} />} />
-      </Switch>
+      <Routes>
+        <Route path={'/:token'} element={<Application />} />
+        <Route path={'/'} element={<Application />} />
+      </Routes>
     </BrowserRouter>
   );
 }
