@@ -1,6 +1,7 @@
 import { ApollonEditor, SVG } from '@ls1intum/apollon';
 import { Action } from 'redux';
 import { Epic, ofType } from 'redux-observable';
+import { Observable } from 'rxjs';
 import { concatAll, map } from 'rxjs/operators';
 import { ApplicationState } from '../../../components/store/application-state';
 import { StopAction, StopActionType } from '../../actions';
@@ -8,7 +9,9 @@ import { DiagramRepository } from '../../diagram/diagram-repository';
 import { FileDownloadAction, FileDownloadActionTypes } from '../../file-download/file-download-types';
 import { ExportActionTypes, ExportPDFAction } from '../export-types';
 
-export const exportPDFEpic: Epic<Action, StopAction | FileDownloadAction, ApplicationState> = (action$, store) => {
+export const exportPDFEpic: Epic<Action, StopAction | FileDownloadAction, ApplicationState> = (
+  action$: Observable<Action<ExportActionTypes>>,
+) => {
   return action$.pipe(
     ofType(ExportActionTypes.EXPORT_PDF),
     map((action) => action as ExportPDFAction),
